% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calinski_harabasz}
\alias{calinski_harabasz}
\title{Calculate Calinski-Harabasz score}
\usage{
calinski_harabasz(df, k = 9, ...)
}
\arguments{
\item{df}{data frame to calculate the Calinski-Harabasz from.}

\item{k}{number of clusters.}

\item{...}{other parameters passed to \code{\link[fpc:calinhara]{fpc::calinhara()}}}
}
\description{
Calculate Calinski-Harabasz score
}
