% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{n_cluster_message}
\alias{n_cluster_message}
\alias{cluster_method_input}
\alias{n_clusters_input}
\alias{cluster_variable_input}
\alias{n_cluster_plot_output}
\alias{cluster_size_bar_plot_output}
\alias{profile_plot_output}
\alias{cluster_pairs_plot_output}
\alias{bivariate_cluster_plot_output}
\alias{discriminant_projection_plot_output}
\alias{dependent_variable_input}
\alias{dependent_variable_plot_output}
\alias{dependent_variable_table_output}
\alias{dependent_null_hypothesis_output}
\alias{optimal_clusters_plot_output}
\alias{cluster_valdiation_plot_output}
\alias{cluster_validation_distribution_plot_output}
\alias{cluster_module}
\title{Output for printing status messages from the Shiny module.}
\usage{
n_cluster_message(id)

cluster_method_input(id)

n_clusters_input(
  id,
  label = "Number of clusters:",
  min = 2,
  max = 10,
  value = 4
)

cluster_variable_input(id)

n_cluster_plot_output(id)

cluster_size_bar_plot_output(id, ...)

profile_plot_output(id, ...)

cluster_pairs_plot_output(id, ...)

bivariate_cluster_plot_output(id, ...)

discriminant_projection_plot_output(id, ...)

dependent_variable_input(id)

dependent_variable_plot_output(id, ...)

dependent_variable_table_output(id)

dependent_null_hypothesis_output(id)

optimal_clusters_plot_output(id)

cluster_valdiation_plot_output(id)

cluster_validation_distribution_plot_output(id)

cluster_module(
  id,
  data,
  default_vars = names(data())[sapply(data(), function(x) {
     is.numeric(x)
 })],
  default_dependent_variable = NULL,
  se_factor = 1
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{label}{label for the slider input.}

\item{min}{The minimum value (inclusive) that can be selected.}

\item{max}{The maximum value (inclusive) that can be selected.}

\item{value}{The initial value of the slider.}

\item{...}{other parmaeters passed to \code{\link[shiny:plotOutput]{shiny::plotOutput()}}.}

\item{data}{a function to return the data (probably a reactive function).}

\item{default_vars}{character list for the variables to include by default.}

\item{default_dependent_variable}{the name of the dependent variable, or NULL for none.}

\item{se_factor}{how many standard errors to plot.}
}
\description{
Table output for analyzing the dependent variable from the clusters. This will
either be an ANOVA for a quantitative dependent variable or a chi-squared test
for qualitative dependent variable.
}
\seealso{
\code{\link[=profile_plot]{profile_plot()}}

\code{\link[GGally:ggpairs]{GGally::ggpairs()}}
}
