% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clav.vanDongen.R
\name{clav.vanDongen}
\alias{clav.vanDongen}
\title{Computes the van Dongen criterion
(2n− ∑_i (max_j n(i,j)) − ∑_j (max_i n(i,j)) )/2n}
\usage{
clav.vanDongen(x, y)
}
\arguments{
\item{x}{A vector with cluster assignments.}

\item{y}{A vector with cluster assignments.}
}
\value{
The value of van Dongen criterion.
}
\description{
Computes the van Dongen criterion
(2n− ∑_i (max_j n(i,j)) − ∑_j (max_i n(i,j)) )/2n
}
\examples{
d<-vegan::vegdist (iris[,3:4], method = "euclidean")
c<-cluster::pam (d, 3, diss = TRUE)
clav.vanDongen(c$clustering, unclass(iris$Species))
}
