% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{cluster_explainer_shiny}
\alias{cluster_explainer_shiny}
\alias{cluster_shiny}
\alias{clav_shiny_ui}
\alias{clav_shiny_server}
\title{Run the Cluster Explainer Shiny Application}
\usage{
cluster_explainer_shiny(...)

cluster_shiny(...)

clav_shiny_ui()

clav_shiny_server(input, output, session)
}
\arguments{
\item{...}{named list of data frames to make available in the Shiny application.}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\description{
Run the Cluster Explainer Shiny Application

Run the Shiny application.

UI for cluster analysis application

Server function for cluster analysis application
}
