% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_agreement_fit.R
\name{cluster_agreement_fit}
\alias{cluster_agreement_fit}
\alias{plot.cluster_agreement_fit}
\alias{hist.cluster_agreement_fit}
\alias{print.cluster_agreement_fit}
\alias{summary.cluster_agreement_fit}
\alias{cluster_agreement}
\title{Cluster Agreement Fit}
\usage{
cluster_agreement_fit(df, k = 2:6, verbose = interactive(), ...)

\method{plot}{cluster_agreement_fit}(x, palette_type = "qual", palette = 1, palette_direction = -1, ...)

\method{hist}{cluster_agreement_fit}(x, exclude_agreements = FALSE, ...)

\method{print}{cluster_agreement_fit}(x, ...)

\method{summary}{cluster_agreement_fit}(object, thresholds, ...)

cluster_agreement(cv)
}
\arguments{
\item{df}{data frame to estimate clusters. Note that all columns will be used in the estimation.}

\item{k}{vector indicating the number of clusters to calculate the agreement fit for.}

\item{verbose}{whether the function should print the status while running.}

\item{...}{currently not used.}

\item{x}{results from \code{\link[=cluster_agreement_fit]{cluster_agreement_fit()}}}

\item{palette_type}{One of "seq" (sequential), "div" (diverging) or "qual" (qualitative)}

\item{palette}{f a string, will use that named palette.}

\item{palette_direction}{Sets the order of colours in the scale.}

\item{exclude_agreements}{if \code{TRUE} agreements of zero or one will be excluded from the bar
plot. This may be helpful for seeing the pattern of how much agreement there may be
when the mode is zero and/or one.}

\item{object}{results from \code{\link[=cluster_agreement_fit]{cluster_agreement_fit()}}}

\item{thresholds}{a vector of thresholds to consider "perfect" agreement. Specifically, values, x,
where x < 0 + threshold or x > (1 - threshold) are considered to agree. Additional columns
of the format \code{fitXXX} where \code{XXX} is the threshold.}

\item{cv}{a \code{\link[=cluster_validation]{cluster_validation()}} object.}
}
\value{
a list of length \code{length(k)}. Each element is a list with the following elements:
\describe{
\item{k}{the number of clusters estimated.}
\item{cv}{the results from \code{\link[=cluster_validation]{cluster_validation()}}}
\item{ca_mat}{the results from \code{\link[=cluster_agreement]{cluster_agreement()}}}
\item{caf}{the cluster agreement fit.}
}

a ggplot2 expression.

a ggplot2 expression.

a data.frame providing the agreement fit for each k.

a data.frame providing the agreement fit for each k.

a matrix with how often each observations cluster membership agrees with all other
observations across all bootstrap sample.
}
\description{
This function calculates the cluster agreement fit for varying cluster sizes (k). Note that this
function will call \code{\link[=cluster_validation]{cluster_validation()}} for each k hence this function may take a while to
execute for large data frames and large k.

This function will calculate how the cluster membership agreement among all observations.
For each bootstrap sample, cluster membership is determined by using the \code{predict()} function
so that each observation has exactly one cluster ID. This also means that both in- and
out-of-bag samples are contributing to the fit statistic. The value in the matrix is a value
between zero and one indicating the proportion of times the two observations had the same
cluster membership.
}
\details{
The cluster agreement fit is a measure of how often observations are either in the same cluster
or not. In a perfect agreement situation the matrix returned by \code{\link[=cluster_agreement]{cluster_agreement()}} will have
all zeros or all ones indicating that a pair of observations are never in the same cluster or
always in the same cluster, respectively.
}
