% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{cluster_module}
\alias{cluster_module}
\title{Shiny module for cluster analysis.}
\usage{
cluster_module(
  id,
  data,
  default_vars = names(data())[sapply(data(), function(x) {
     is.numeric(x)
 })],
  default_dependent_variable = NULL,
  se_factor = 1
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data}{a function to return the data (probably a reactive function).}

\item{default_vars}{character list for the variables to include by default.}

\item{se_factor}{how many standard errors to plot.}

\item{default_dependnet_variable}{the name of the dependent variable, or NULL for none.}
}
\description{
Shiny module for cluster analysis.
}
