% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{cluster_size_bar_plot_output}
\alias{cluster_size_bar_plot_output}
\title{Plot output for the bar plot of cluster sizes.}
\usage{
cluster_size_bar_plot_output(id, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{...}{other parameters passed to \code{\link[shiny:plotOutput]{shiny::plotOutput()}}}
}
\description{
Plot output for the bar plot of cluster sizes.
}
