% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_validation.R,
%   R/cluster_validation_plot.R, R/cluster_validation_print.R,
%   R/cluster_validation_summary.R
\name{cluster_validation}
\alias{cluster_validation}
\alias{plot.clustervalidation}
\alias{print.clustervalidation}
\alias{summary.clustervalidation}
\alias{plot.cv_summary}
\title{Cluster profile validation}
\usage{
cluster_validation(
  df,
  n_clusters = 2,
  cluster_fun = kmeans,
  get_cluster_fun = function(x) {
     x$cluster
 },
  oob_predict_fun = function(fit, newdata) {
     predict(fit, newdata = newdata)
 },
  summary_fun = mean,
  n_samples = 100,
  sample_size = 0.5 * nrow(df),
  replace = FALSE,
  standardize = TRUE,
  seed,
  verbose = interactive(),
  ...
)

\method{plot}{clustervalidation}(
  x,
  facet = TRUE,
  plot_complete = TRUE,
  plot_oob = TRUE,
  point_alpha = 0.1,
  point_size = 1,
  line_alpha = 0.1,
  line_width = 1,
  xlab = "",
  ylab = ifelse(attr(x, "standardize"), "Mean Standard Score", "Mean Score"),
  ...
)

\method{print}{clustervalidation}(x, ...)

\method{summary}{clustervalidation}(object, in_sample = FALSE, oob_sample = TRUE, ...)

\method{plot}{cv_summary}(x, ...)
}
\arguments{
\item{df}{data frame to estimate clusters. Note that all columns will be used
in the estimation.}

\item{n_clusters}{the number of clusters to estimate.}

\item{cluster_fun}{the function used to estimate the clusters.}

\item{get_cluster_fun}{the function used to get the cluster classes. This function
takes one parameter, the result of \code{cluster_fun}.}

\item{oob_predict_fun}{the function used to get predictions from the out-of-bag
sample. Function takes two parameters, the first is the results of
\code{cluster_fun}, the second is the  out-of-bag sample data frame.}

\item{summary_fun}{the function used to calculate the statistic for each
cluster and iteration. Defaults to \code{mean}.}

\item{n_samples}{the number of random samples to draw.}

\item{sample_size}{the size of each random sample. Defaults to 50\% of observations.}

\item{replace}{whether sampling should be done with replacement.}

\item{standardize}{whether the variables should be standardized before
estimating clusters.}

\item{seed}{random number seed. Note that the seed is set before each iteration
to \code{seed + i} where \code{i} is the iteration number.}

\item{verbose}{whether the function should print the status while running.}

\item{...}{currently not used.}

\item{x}{result of \code{summary.clustervalidation()}}

\item{facet}{whether the different samples (i.e. in sample, out-of-bag, and
and complete) should be faceted.}

\item{plot_complete}{whether the profile line using the complete data set
should be plotted.}

\item{plot_oob}{whether the out-of-bag sample should be plotted.}

\item{point_alpha}{the alpha (transparency) level for points.}

\item{line_alpha}{the alpha (transparency) level for lines.}

\item{xlab}{label for the x-axis.}

\item{ylab}{label for the y-axis.}

\item{object}{the results of \code{cluster_validation}.}

\item{in_sample}{plot the in sample results.}

\item{oob_sample}{plot the out-of-bag sample results.}
}
\value{
a list with the following elements:
\describe{
\item{complete_sample}{data frame of results using the entire data set.}
\item{in_sample}{data frame of in sample results.}
\item{oob_sample}{data frame of out-of-bag results.}
\item{complete_model_fit}{model fit for the full data set.}
\item{in_sample_model_fits}{model fits for each sample.}
}
Each of these data frames contain four columns:
\describe{
\item{iter}{the iteration}
\item{cluster}{the cluster}
\item{variable}{the variable}
\item{value}{the mean value of \code{variable} for the given cluster and interation}
}

a ggplot2 expression.

a ggplot2 expression.
}
\description{
This function takes multiple random samples from the provided data frame,
estimate cluster membership, and calculate the mean (the default, a different
statistic can be specified using the \code{summary_fun} parameter, e.g. \code{median}
may be appropriate) for each cluster using both the in sample and out-of-bag
(i.e. out of sample). For the out-of-bag sample
}
\details{
The number of rows in the resulting data frames will be equal to:
\code{ncol(df) * n_samples * n_clusters}.
}
