% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{cluster_variable_input}
\alias{cluster_variable_input}
\title{Input to select the variables to perform the cluster analysis with.}
\usage{
cluster_variable_input(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
Input to select the variables to perform the cluster analysis with.
}
