% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_fit_stats.R
\docType{data}
\name{clustering_fit_stats}
\alias{clustering_fit_stats}
\title{Description of fit statistics for cluster analysis}
\format{
A data frame with 6 rows and 5 variables
}
\usage{
data(clustering_fit_stats)
}
\description{
This data frame contains descriptions of the fit statistics used by the \code{\link[=optimal_clusters]{optimal_clusters()}}
function. It contains the following columns:
}
\details{
\describe{
\item{Fit_Statistic}{The name of the fit statistic.}
\item{Reference}{Reference to first publication.}
\item{Description}{A brief description of the fit statistic.}
\item{Desired_Outcome}{How to interpret the statistic (i.e. what is the desired outcome).}
\item{R_Function}{The R function used to get the statistic. Note that \code{clav} wraps most of these functions.}
}
}
\keyword{datasets}
