% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{dependent_null_hypothesis_output}
\alias{dependent_null_hypothesis_output}
\title{Output from the dependent varaible analysis.}
\usage{
dependent_null_hypothesis_output(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
Output from the dependent varaible analysis.
}
