% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{dependent_variable_input}
\alias{dependent_variable_input}
\title{Shiny input to select the dependent (outcome) variable.}
\usage{
dependent_variable_input(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
Shiny input to select the dependent (outcome) variable.
}
