% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{dependent_variable_plot_output}
\alias{dependent_variable_plot_output}
\title{TODO: REMOVE}
\usage{
dependent_variable_plot_output(id, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
TODO: REMOVE
}
