% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{dependent_variable_table_output}
\alias{dependent_variable_table_output}
\title{Output of the dependent variable analysis.}
\usage{
dependent_variable_table_output(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
Table output for analyzing the dependent variable from the clusters. This will
either be an ANOVA for a quantitative dependent variable or a chi-squared test
for qualitative dependent variable.
}
