% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ev.vanDongen.R
\name{ev.vanDongen}
\alias{ev.vanDongen}
\title{Computes the van Dongen criterion}
\usage{
ev.vanDongen(x, y)
}
\arguments{
\item{x}{A vector with cluster assignments.}

\item{y}{A vector with cluster assignments.}
}
\value{
The value of van Dongen criterion. An integer between 0 and 1. Lower values indicate better cluster similarity.
}
\description{
Computes the van Dongen criterion
}
\examples{
d<-vegan::vegdist (iris[,3:4], method = "euclidean")
c<-cluster::pam (d, 3, diss = TRUE)
ev.vanDongen(c$clustering, unclass(iris$Species))
}
