% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_clusters.R, R/optimal_clusters_plot.R,
%   R/utils.R
\name{optimal_clusters}
\alias{optimal_clusters}
\alias{plot.optimalclusters}
\alias{print.optimalclustersplot}
\alias{wss}
\alias{silhouette_score}
\alias{calinski_harabasz}
\alias{davies_bouldin}
\alias{rand_index}
\title{Utility function used to determine the optimal number of clusters.}
\usage{
optimal_clusters(
  df,
  max_k = 9,
  cluster_fun = stats::kmeans,
  wss = TRUE,
  silhoutte = TRUE,
  gap = TRUE,
  calinski_harabasz = TRUE,
  davies_bouldin = TRUE,
  rand_index = TRUE
)

\method{plot}{optimalclusters}(x, ...)

\method{print}{optimalclustersplot}(x, ...)

wss(df, k = 9, cluster_fun = stats::kmeans)

silhouette_score(df, k = 9, cluster_fun = stats::kmeans, ...)

calinski_harabasz(df, k = 9, cluster_fun = stats::kmeans, ...)

davies_bouldin(df, k = 9, cluster_fun = stats::kmeans, ...)

rand_index(df, k = 9, cluster_fun = stats::kmeans, ...)
}
\arguments{
\item{df}{data frame to calculate the Rand index from.}

\item{max_k}{maximum number of clusters to estimate.}

\item{cluster_fun}{clustering function.}

\item{wss}{whether to calculate the within sum of squares.}

\item{silhoutte}{whether to calculate the Silhoutte index.}

\item{gap}{whether to calculate the Gap statistic.}

\item{calinski_harabasz}{whether to calculate the Calinski Harabasz index}

\item{davies_bouldin}{whether to calculate the Davies Bouldin index.}

\item{rand_index}{whether to calculate the Rand index.}

\item{x}{results from \code{plot.optimalclusters}}

\item{...}{other parameters passed to \code{\link[fossil:rand.index]{fossil::rand.index()}}.}

\item{k}{number of clusters.}
}
\value{
a data frame with various metrics used to determine the optimal number
of clusters. Each row corresponds to k ranging from 1 to \code{max_k} and has the
following columns (if the corresponding metric was requested):
\describe{
\item{k}{Number of clusters}
\item{wss}{Within sum of squares}
\item{silhoutte}{Silhouette analysis measures the quality of clustering and provides an insight
into the separation distance between the resulting clusters. A higher
silhouette score indicates that the object is well matched to its own cluster
and poorly matched to neighboring clusters.}
\item{gap}{The Gap Statistic compares the total within intra-cluster variation for different
values of k with their expected values under null reference distribution of the data.
The optimal k is the value that yields the largest gap statistic.}
\item{calinski_harabasz}{This Index evaluates clusters based on their compactness and separation.
The index is calculated using the ratio of between-cluster variance to
within-cluster variance, with higher values indicate better-defined clusters}
\item{davies_bouldin}{The Davies-Bouldin index is a measure of the how much separation there is
between clusters. Lower values of the Davies-Bouldin index indicate a
model with better separation.}
\item{rand_index}{It calculates the proportion of agreement between the two clusters, considering
both the pairs of elements that are correctly assigned to the same or different
clusters. Higher values indicate greater similarity and better clustering quality.}
}

a list of ggplot2 figures for each fit statistic.

displays a grid of plots.
}
\description{
This function will calculate a number of metrics used to determine the optimal number of clusters.
The result will be a data frame where each row corresponds to the number of clusters (k) and
the columns are the various metrics. An S3 plotting function is provided.
}
