% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust2.R
\name{hclust2}
\alias{hclust2}
\title{Hierarchical Clustering}
\usage{
hclust2(x, k, ...)
}
\arguments{
\item{x}{data frame of variables to find clusters for.}

\item{k}{the desired number of clusters.}

\item{...}{other parameters passed to \code{\link[stats:hclust]{stats::hclust()}}.}
}
\value{
the results of \code{\link[stats:hclust]{stats::hclust()}} along with additional values, namely:
\describe{
\item{k}{the desired number of clusters.}
\item{cluster}{a vector with the cluster membership.}
}
}
\description{
This is a wrapper function to \code{\link[stats:hclust]{stats::hclust()}} to coerce the call and results to conform
to the same structure as \code{\link[stats:kmeans]{stats::kmeans()}}.
}
