% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.xb.R
\name{iv.xb}
\alias{iv.xb}
\title{Computes the Xie-Beni index}
\usage{
iv.xb(clustering, centers, diss, variant = "weighted")
}
\arguments{
\item{clustering}{A vector with cluster assignments or a membership matrix.}

\item{centers}{A vector with cluster center IDs.}

\item{diss}{A distance matrix.}

\item{variant}{Weighted (XB*) or normal (XB) formula. Check the reference for more detail.}
}
\value{
A number.
}
\description{
The Xie-Beni index is an internal cluster validation index. It deﬁnes the intercluster separation as the minimum square distance between cluster centers, and the intraclus ter compactness as the mean square distance between each data object and its cluster center. It can be used for fuzzy clustering as well. The weighted variant (default is usually the better option.
}
\examples{
d<-vegan::vegdist (iris[,3:4], method = "euclidean")
models = list()

for (k in 3:5) 
  models[[length(models)+1]] = fclust::FKM.med(as.matrix(d), k)

for (v in c('normal','weighted'))
  for (f in models)
    print(paste (ncol(f$U),'clusters -', v, '-', iv.xb(f$U, f$medoid, d, variant=v)))
}
\references{
\url{https://www.researchgate.net/profile/Minho_Kim7/publication/222750120_New_indices_for_cluster_validity_assessment/links/02e7e53c567697580d000000.pdf}
}
