% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_iterative.R
\name{kmeans_iterative}
\alias{kmeans_iterative}
\title{Function to perform K-means step by step and save iterations}
\usage{
kmeans_iterative(x, centers, max_iter = 50)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix (such as a
numeric vector or a data frame with all numeric columns).}

\item{centers}{either a numeric value for k or a k by ncol(x) matrix of the initial centers.}

\item{max_iter}{maximum number of iterations before stopping if convergence doesn't occur first.}
}
\value{
a list where each element corresponds to an interation in the k-means algorithm. For each
element there are three values: \code{iteration} (numeric for the which iteration), \code{centers}
(matrix for the cluster centers), and \code{clusters} (vector of cluster assignment for all
observations in \code{x}).
}
\description{
This function performs K-means clustering but saves the cluster assignment and centers for each
iteration of the algorithm.
}
