% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{n_cluster_message}
\alias{n_cluster_message}
\title{Output for printing status messages from the Shiny module.}
\usage{
n_cluster_message(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
Output for printing status messages from the Shiny module.
}
