% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{n_cluster_plot_output}
\alias{n_cluster_plot_output}
\title{Plot output for the figure to determine the optimal number of clusters.}
\usage{
n_cluster_plot_output(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
Plot output for the figure to determine the optimal number of clusters.
}
