% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{n_clusters_input}
\alias{n_clusters_input}
\title{Slider input for the desired number of clusters.}
\usage{
n_clusters_input(
  id,
  label = "Number of clusters:",
  min = 2,
  max = 10,
  value = 4
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{label}{label for the slider input.}

\item{min}{The minimum value (inclusive) that can be selected.}

\item{max}{The maximum value (inclusive) that can be selected.}

\item{value}{The initial value of the slider.}
}
\description{
Slider input for the desired number of clusters.
}
