% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_clusters.R, R/optimal_clusters_plot.R
\name{optimal_clusters}
\alias{optimal_clusters}
\alias{plot.optimalclusters}
\title{Utility function used to determine the optimal number of clusters.}
\usage{
optimal_clusters(df, max_k = 9)

\method{plot}{optimalclusters}(x, ...)
}
\arguments{
\item{df}{data frame to determine the optimal number of clusters from.}

\item{max_k}{maximun number of clusters to estimate.}

\item{x}{the result of \code{\link[=optimal_clusters]{optimal_clusters()}}}

\item{...}{other parameters passed to \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}}
}
\value{
a data frame with various metrics used to determine the optimal number
of clusters. Each row corresponds to k ranging from 1 to \code{max_k} and has the
following columns:
\describe{
\item{k}{number of clusters}
\item{wss}{within sum of squares}
\item{silhoutte}{}
\item{gap}{}
\item{calinski_harabasz}{}
\item{davies_bouldin}{}
}

a ggplot2 expression
}
\description{
Utility function used to determine the optimal number of clusters.
}
