% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_fit.R
\name{overlap_fit}
\alias{overlap_fit}
\alias{plot.overlap}
\title{Cluster analysis fit statistic using bootstrap distributions}
\usage{
overlap_fit(df, k = 2:6, verbose = interactive(), ...)

\method{plot}{overlap}(x, se_factor = 1, ...)
}
\arguments{
\item{df}{data frame containing the variables to use for cluster.}

\item{k}{an integer vector for the cluster sizes to estimate the overlap for.}

\item{verbose}{whether to print status as the boostrap samples are estimated.}

\item{...}{currently not used.}

\item{x}{results from \code{\link[=overlap_fit]{overlap_fit()}}.}

\item{se_factor}{factor to multiple the standard error. For example, for a 95\% confidence interval
set \code{se_factor = 1.96}.}
}
\value{
an object that inherits from a data.frame with the following columns:
\describe{
\item{variable}{the variable name.}
\item{C1}{the label for cluster 1.}
\item{C2}{the label for cluster 2.}
\item{overlap}{how much of the bootstrap distributions overlap from the distribution of cluster
centers for clusters 1 and 2 for \code{variable}.}
\item{k}{the number of clusters.}
}
There is a custom \code{plot()} function for the returned object.

a \code{ggplot2} expression.
}
\description{
Note: This will always use bootstrapping.
}
\examples{
\dontrun{
data(daacs, package='clav')
cluster_vars <- c("Motivation", "Metacognition", "Strategies", "Mathematics", "Reading", "Writing")
overlap_fit <- overlap_fit(daacs[,cluster_vars])
plot(overlap_fit)
}
}
