% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa2015.R
\docType{data}
\name{pisa2015}
\alias{pisa2015}
\title{Student questionnaire and science performance data from the 2015 PISA}
\format{
A data frame with 20,571 rows and 8 variables
}
\source{
https://www.oecd.org/en/data/datasets/pisa-2015-database.html
}
\usage{
data(pisa2015)
}
\description{
This is a subset of the Programme of International Student Assessment (PISA)
including student questionnaire data and their performance on the science
assessment. Data from the United States and Canada are included.
}
\details{
There are seven variables:
\itemize{
\item \code{country.} The country the student is from (USA or Canada).
\item \code{science_score.} Average science score over the 10 plausible values.
\item \code{principals.} Average scale score for students self-reported understanding of scientific principals.
\item \code{belonging.} Average scale score for belonging.
\item \code{interest.} Average scale score for interest in science.
\item \code{enjoyment.} Average scale score for enjoyment doing science.
\item \code{motivation.} Average scale score for motivation in science.
\item \code{efficacy.} Average scale score for students self-efficacy in science.
}

All numeric scores have been converted to standard scores (z-scores).

Below are the Likert response questions students completed that are used to
calculate the belonging, interest, motivation, efficacy, and principals
variables. The values in parentheses correspond to the variable names in the
original PISA database. The questionnaire can be downloaded from
\href{https://www.oecd.org/content/dam/oecd/en/data/datasets/pisa/pisa-2015-datasets/questionnaires/Student\%20questionnaire\%20PBA\%20main\%20PISA\%202015.pdf}{the OECD PISA website}.

\strong{Belonging:} Thinking about your school: to what extent do you agree with the following statements? (\code{ST034})
\itemize{
\item I feel like an outsider (or left out of things) at school.
\item I make friends easily at school.
\item I feel like I belong at school.
\item I feel awkward and out of place in my school.
\item Other students seem to like me.
\item I feel lonely at school.
}

\strong{Interest:} How much do you disagree or agree with the statements about yourself below? (\code{ST094})
\itemize{
\item I generally have fun when I am learning \emph{broad science} topics.
\item I like reading about \emph{broad science}.
\item I am happy working on \emph{broad science} topics.
\item I enjoy acquiring new knowledge in \emph{broad science}.
\item I am interested in learning about \emph{broad science}.
}

\strong{Enjoyment:} How much do you disagree or agree with the statements about yourself below? (\code{ST094})
\itemize{
\item I generally have fun when I am learning \emph{broad science} topics.
\item I like reading about \emph{broad science}.
\item I am happy working on \emph{broad science} topics.
\item I enjoy acquiring new knowledge in \emph{broad science}.
\item I am interested in learning about \emph{broad science}.
}

\strong{Motivation:} How much do you agree with the statements below? (\code{ST113})
\itemize{
\item Making an effort in my \emph{school science} subject(s) is worth it because this will help me in the work I want to do later on.
\item What I learn in my \emph{school science} subject(s) is important for me because I need this for what I want to do later on.
\item Studying my \emph{school science} subject(s) is worthwhile for me because what I learn will improve my career prospects.
\item Many things I learn in my \emph{school science} subject(s) will help me to get a job.
}

\strong{Efficacy:} How easy do you think it would be for you to perform the following tasks on your own? (\code{ST129})
\itemize{
\item Recognise the science question that underlies a newspaper report on a health issue.
\item Explain why earthquakes occur more frequently in some areas than in others.
\item Describe the role of antibiotics in the treatment of disease.
\item Identify the science question associated with the disposal of garbage.
\item Predict how changes to an environment will affect the survival of certain species.
\item Interpret the scientific information provided on the labelling of food items.
\item Discuss how new evidence can lead you to change your understanding about the possibility of life on Mars.
\item Identify the better of two explanations for the formation of acid rain.
}

\strong{Principals:} How much do you disagree or agree with the statements below? (\code{ST131})
\itemize{
\item A good way to know if something is true is to do an experiment.
\item Ideas in \emph{broad science} sometimes change.
\item Good answers are based on evidence from many different experiments.
\item It is good to try experiments more than once to make sure of your findings.
\item Sometimes \emph{broad science} scientists change their minds about what is true in science.
\item The ideas in \emph{broad science} science books sometimes change.
}
}
\keyword{datasets}
