% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict.kmeans}
\alias{predict.kmeans}
\title{Predict cluster class for k-means cluster}
\usage{
\method{predict}{kmeans}(object, newdata, method = c("classes", "centers"), ...)
}
\arguments{
\item{object}{results of \code{\link[stats:kmeans]{stats::kmeans()}}.}

\item{newdata}{data frame to make class predictions.}

\item{method}{method used to make predictions.}

\item{...}{currently not used.}
}
\value{
vector of predicted clusters.
}
\description{
Predict cluster class for k-means cluster
}
