% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_p_value}
\alias{print_p_value}
\title{Utility function to print p-values}
\usage{
print_p_value(p_value, digits = 2)
}
\arguments{
\item{p_value}{the p-value.}

\item{digits}{number of digits to print.}
}
\value{
a character version of the p-value.
}
\description{
This function will print a pretty p-value. If the p-value is less than
10^(-digits), for example 0.01 with digits = 2, then the result will be
"p < 0.01".
}
