% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{profile_plot}
\alias{profile_plot}
\title{Profile plot for cluster analysis.}
\usage{
profile_plot(
  df,
  clusters,
  df_dep,
  standardize = TRUE,
  bonferroni = TRUE,
  label_means = TRUE,
  label_profile_means = label_means,
  label_outcome_means = label_means,
  center_band = 0.25,
  center_fill = "#f0f9e8",
  center_alpha = 0.1,
  text_size = 4,
  hjust = 0.5,
  point_size = 2,
  se_factor = 1.96,
  color_palette = 2,
  cluster_labels,
  cluster_order,
  label_clusters = TRUE,
  cluster_label_x,
  cluster_label_hjust = 5,
  ylab = ifelse(standardize, "Mean Standard Score", "Mean Score"),
  title = "Cluster Profiles"
)
}
\arguments{
\item{df}{data.frame with the columns used for the cluster analysis.}

\item{clusters}{vector indicating what cluster each row of \code{df} belongs to.}

\item{df_dep}{a data.frame with any dependent variables to include in the plot (optional).}

\item{standardize}{if TRUE values in \code{df} will be converted to z-scores.}

\item{bonferroni}{if TRUE Bonferroni adjusted error bars will be plotted.}

\item{label_means}{label the mean values of both clusters and outcome variables.}

\item{label_profile_means}{label the mean values of the clusters.}

\item{label_outcome_means}{label the mean values of outcome variables.}

\item{center_band}{the percentage around the mean to shade. This only works
if \code{standardize = TRUE}.}

\item{center_fill}{the color of the center band.}

\item{center_alpha}{the transparency level of the center band.}

\item{text_size}{text size.}

\item{hjust}{horizontal adjustment of labels.}

\item{point_size}{size of points passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{se_factor}{critical value used ot determine the width of standard error bars.}

\item{color_palette}{the color palette to use. See \code{ggplot2:scale_color_brewer()}
for more details.}

\item{cluster_labels}{labels for the clusters.}

\item{cluster_order}{order of clusters on the x-axis.}

\item{label_clusters}{whether to label clusters on the main panel.}

\item{cluster_label_x}{cluster labels.}

\item{cluster_label_hjust}{horizontal adjustment for y-axis labels.}

\item{ylab}{label for the y-axis.}

\item{title}{plot title.}
}
\description{
Profile plot for cluster analysis.
}
