% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_this}
\alias{scale_this}
\title{Convert vector to z-scores}
\usage{
scale_this(x)
}
\arguments{
\item{x}{vector to convert.}
}
\value{
vector of standard scores.
}
\description{
Convert vector to z-scores
}
