% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_by.R
\name{se}
\alias{se}
\title{Calculate the standard error of the mean}
\usage{
se(x, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{...}{other parameters passed to \code{sd}.}
}
\value{
the standard error.
}
\description{
Calculate the standard error of the mean
}
