% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{silhouette_score}
\alias{silhouette_score}
\title{Calculate Silhouette score}
\usage{
silhouette_score(df, k = 9, ...)
}
\arguments{
\item{df}{data frame to calculate the Silhouette from.}

\item{k}{number of clusters.}

\item{...}{other parameters passed to \code{\link[cluster:silhouette]{cluster::silhouette()}}}
}
\description{
Calculate Silhouette score
}
