% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wss}
\alias{wss}
\title{Calculate within sum of squares.}
\usage{
wss(df, k = 9)
}
\arguments{
\item{df}{data frame to calculate the WSS from.}

\item{k}{number of clusters.}
}
\description{
Calculate within sum of squares.
}
