% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{dependent_variable_plot_output}
\alias{dependent_variable_plot_output}
\title{Shiny output for the dependnet sample plot.}
\usage{
dependent_variable_plot_output(id, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{...}{other parmaeters passed to \code{\link[shiny:plotOutput]{shiny::plotOutput()}}.}
}
\description{
Shiny output for the dependnet sample plot.
}
