% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_by.R
\name{describe_by}
\alias{describe_by}
\title{Basic summary statistics by group}
\usage{
describe_by(
  df,
  group,
  group_name = "group",
  funs = list(n = length, mean = mean, sd = sd, se = se, median = median, min = min, max
    = max, range = function(x) {
     diff(range(x))
 })
)
}
\arguments{
\item{df}{a data frame to summarize.}

\item{group}{a grouping vector.}

\item{funs}{a named list of functions. The names of the list will be used as the column names
in the returned data frame.}
}
\value{
a data frame with a column \code{group} corresponding to the group variable and column
\code{variable} corresponding to the names of the data frame. Additional columns for each
function in the \code{funs} parameter.
}
\description{
Calculate basic summary statistics by a grouping variable. This function is inspired by the
\code{psych::describeBy()} function but has been implemented to always return a data frame using
the tidyverse. The \code{funs} parameter provides additional flexibility regarding what summary
statistics are calculated.
}
\examples{
data(pisa2015, package = 'clav')
describe_by(df = pisa2015[,c('science_score', 'efficacy', 'motivation')], group = pisa2015$country)
}
