% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ev.vanDongen.R
\name{ev.vanDongen}
\alias{ev.vanDongen}
\title{Computes the van Dongen criterion}
\usage{
ev.vanDongen(x, y, normalized = F)
}
\arguments{
\item{x}{A vector with cluster assignments.}

\item{y}{A vector with cluster assignments.}

\item{normalized}{T or F(default). Indicate whether to normalize the measure or not.}
}
\value{
A number between 0 and 1.
}
\description{
The van Dongen criterion is an external validation measure of the representativeness of the majority objects in each class and each cluster. It is an number between 0 and 1 and lower values indicate better cluster similarity. Even though the van Dongen measure is normalized, this is achieved by deleting by 2*n which is an absolute upper bound that is not tight. Hence, a special normalization with a tighter upper bound is also provided (by using \code{normalized=T}).
}
\examples{
d<-vegan::vegdist (iris[,3:4], method = "euclidean")
c<-cluster::pam (d, 3, diss = TRUE)
ev.vanDongen(c$clustering, unclass(iris$Species))
}
\references{
\url{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.26.9783&rep=rep1&type=pdf}
}
