% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ev.vanDongen.R
\name{ev.vanDongen}
\alias{ev.vanDongen}
\title{Computes the van Dongen criterion
The van Dongen criterion is an external validation measure of the representativeness of the majority objects in each class and each cluster. It is an number between 0 and 1 and lower values indicate better cluster similarity.}
\usage{
ev.vanDongen(x, y)
}
\arguments{
\item{x}{A vector with cluster assignments.}

\item{y}{A vector with cluster assignments.}
}
\value{
A number between 0 and 1.
}
\examples{
d<-vegan::vegdist (iris[,3:4], method = "euclidean")
c<-cluster::pam (d, 3, diss = TRUE)
ev.vanDongen(c$clustering, unclass(iris$Species))
}
